Check parameter/return value types on shorty signature bindings
When there was only one shorty match method (on a found class), the
match was performed, disregarding the prototypes. This leaded to issue
because some methods were not the one exported (in particular methods
mymethod$number). The context macth then tried to bind illegal classes.
Now there is a layer to check that Android Platform/Java classes are
correct. This is a quick win (full implementation would require too much
time to compute so we are only based on package name).

This lower the rate of false positive detected, indeed.